requirements.txt
aniso8601==8.0.0
asn1crypto==0.24.0
beautifulsoup4==4.7.1
blinker==1.4
bs4==0.0.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
crypto==1.4.1
cryptography==1.9
Django==2.2.1
Flask==1.0.2
Flask-Mail==0.9.1
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
idna==2.8
image==1.5.27
inflect==2.1.0
itsdangerous==1.1.0
Jinja2==2.10.1
lxml==4.3.3
MarkupSafe==1.1.1
mysql==0.0.2
mysqlclient==1.4.2.post1
Naked==0.1.31
Pillow==6.0.0
pycparser==2.19
pycryptodome==3.8.1
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.1
PyYAML==5.1
qrcode==6.1
redis==3.2.1
requests==2.21.0
shellescape==3.4.1
six==1.12.0
soupsieve==1.9.1
sqlacodegen==2.0.1
SQLAlchemy==1.3.3
sqlparse==0.3.0
urllib3==1.24.3
Werkzeug==0.15.2
